# openapi.yaml
openapi: 3.0.0
info:
  title: 'PHP Fake User Avatar Api'
  version: '1'
paths:
  /api/login:
    post:
      summary: Login
      description: 'Login a user and generate JWT token'
      operationId: c7798e20b4c32275b766e54e334f3c81
      requestBody:
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                password:
                  type: string
              type: object
              example:
                email: mohammad@joubeh.com
                password: '123456'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  token: { type: string }
                type: object
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTZ9.SLZRXQeFa64LGhAetQ-accbiNNa20bPaWoIbV9T0jx4
        '400':
          description: 'Invalid email or password'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                type: object
              example:
                message: 'email or password is not valid'
  /api/register:
    post:
      summary: Register
      description: 'Register a user and generate JWT token'
      operationId: e206a767e0d89b11fe2795ff1db52f90
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
              type: object
              example:
                name: mohammad
                email: mohammad@joubeh.com
                password: '123456'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  token: { type: string }
                type: object
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTZ9.SLZRXQeFa64LGhAetQ-accbiNNa20bPaWoIbV9T0jx4
        '400':
          description: 'Email already exists'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                type: object
              example:
                message: 'email already exists'
...